Revert "Add `ETHAddress` field"
This reverts commit 0046dc97e3a9f856ed87c37a06befe4939c7c18f.
