+++ starknet_py/net/full_node_client.py
@@ -37,6 +37,7 @@ from starknet_py.net.client_models import (
     TransactionTrace,
 )
 from starknet_py.net.client_utils import (
+    _create_broadcasted_txn,
     _is_valid_eth_address,
     _to_rpc_felt,
     _to_storage_key,
@@ -49,7 +50,6 @@ from starknet_py.net.models.transaction import (
     DeployAccount,
     Invoke,
 )
-from starknet_py.net.schemas.broadcasted_txn import _create_broadcasted_txn
 from starknet_py.net.schemas.rpc.block import (
     BlockHashAndNumberSchema,
     BlockStateUpdateSchema,
