+++ starknet_py/net/schemas/rpc/transactions.py
@@ -27,7 +27,6 @@ from starknet_py.net.client_models import (
 )
 from starknet_py.net.schemas.common import (
     DAModeField,
-    ETHAddress,
     ExecutionStatusField,
     Felt,
     FinalityStatusField,
@@ -46,7 +45,7 @@ from starknet_py.utils.schema import Schema
 
 class L2toL1MessageSchema(Schema):
     l2_address = Felt(data_key="from_address", required=True)
-    l1_address = ETHAddress(data_key="to_address", required=True)
+    l1_address = Felt(data_key="to_address", required=True)
     payload = fields.List(Felt(), data_key="payload", required=True)
 
     @post_load
